updated tests and feedback from pr
