+++ network-api/networkapi/wagtailpages/pagemodels/buyersguide/products.py
@@ -37,7 +37,9 @@ from wagtail_localize.fields import SynchronizedField, TranslatableField
 
 from networkapi.utility import orderables
 from networkapi.wagtailpages.fields import ExtendedYesNoField
-from networkapi.wagtailpages.forms import BuyersGuideProductCategoryForm
+from networkapi.wagtailpages.pagemodels.buyersguide.forms import (
+    BuyersGuideProductCategoryForm,
+)
 from networkapi.wagtailpages.pagemodels.buyersguide.utils import (
     get_buyersguide_featured_cta,
     get_categories_for_locale,
