+++ network-api/networkapi/wagtailpages/tests/buyersguide/test_article_page.py
@@ -31,8 +31,6 @@ class BuyersGuideArticlePageTest(test_base.WagtailpagesTestCase):
         cls.content_index = buyersguide_factories.BuyersGuideEditorialContentIndexPageFactory(
             parent=cls.buyersguide_homepage,
         )
-        edit_handler = pagemodels.BuyersGuideArticlePage.get_edit_handler()
-        cls.article_page_form = edit_handler.get_form_class()
 
     def test_parents(self):
         self.assertAllowedParentPageTypes(
@@ -210,107 +208,3 @@ class BuyersGuideArticlePageTest(test_base.WagtailpagesTestCase):
         result = article_page.get_secondary_related_articles()
 
         self.assertListEqual(result, [])
-
-    @staticmethod
-    def generate_form_data(article_page, data: dict) -> dict:
-        """
-        Generate valid form data for the BuyersGuide article page form.
-
-        We add the extra fields below to the dict as all these fields are set as
-        required, and the form will return a validation error against them if left out.
-
-        Since we have the "body" field set as required, we need to submit at least
-        one StreamField in the form, or else the form will return a validation error.
-        """
-        return nested_form_data(
-            {
-                **data,
-                "title": article_page.title,
-                "slug": article_page.slug,
-                "body": streamfield([("paragraph", rich_text("Hello world!"))]),
-            }
-        )
-
-    def test_article_page_requires_search_image(self):
-        """
-        Test that the buyersguide article page form will
-        return a validation error if no "search_image" is set.
-        """
-        article_page = buyersguide_factories.BuyersGuideArticlePageFactory(parent=self.content_index)
-
-        test_form = self.article_page_form(
-            data=self.generate_form_data(
-                article_page,
-                {
-                    "search_description": article_page.search_description,
-                    "search_image": None,
-                },
-            ),
-        )
-        self.assertEqual(1, len(test_form.errors))
-        self.assertIn("search_image", test_form.errors)
-        self.assertIn("This field is required.", test_form.errors["search_image"])
-
-    def test_article_page_requires_search_description(self):
-        """
-        Test that the buyersguide article page form will
-        return a validation error if no "search_description" is set.
-        """
-        article_page = buyersguide_factories.BuyersGuideArticlePageFactory(parent=self.content_index)
-
-        test_form = self.article_page_form(
-            data=self.generate_form_data(
-                article_page,
-                {
-                    "search_description": None,
-                    "search_image": article_page.search_image,
-                },
-            ),
-        )
-
-        self.assertEqual(1, len(test_form.errors))
-        self.assertIn("search_description", test_form.errors)
-        self.assertIn("This field is required.", test_form.errors["search_description"])
-
-    def test_article_page_requires_both_search_fields(self):
-        """
-        Test that the buyersguide article page form will return validation
-        errors for both "search_image" and "search_description" fields
-        if neither are updated.
-        """
-        article_page = buyersguide_factories.BuyersGuideArticlePageFactory(parent=self.content_index)
-
-        test_form = self.article_page_form(
-            data=self.generate_form_data(
-                article_page,
-                {
-                    "search_description": None,
-                    "search_image": None,
-                },
-            ),
-        )
-
-        self.assertEqual(2, len(test_form.errors))
-        self.assertIn("search_image", test_form.errors)
-        self.assertIn("search_description", test_form.errors)
-        self.assertIn("This field is required.", test_form.errors["search_image"])
-        self.assertIn("This field is required.", test_form.errors["search_description"])
-
-    def test_article_page_with_search_fields_is_valid(self):
-        """
-        Test that a buyersguide article page form with
-        the search fields set is valid with no errors.
-        """
-        article_page = buyersguide_factories.BuyersGuideArticlePageFactory(parent=self.content_index)
-
-        test_form = self.article_page_form(
-            data=self.generate_form_data(
-                article_page,
-                {
-                    "search_description": article_page.search_description,
-                    "search_image": article_page.search_image,
-                },
-            ),
-        )
-
-        self.assertEqual(0, len(test_form.errors))
